Do not alter WCS on saving to MongoDB
This hasn't helped problems with sources being displayed incorrectly in
Banana: we will fix it in the Banana codebase instead.

I'm retaining the code that does the munging because I imagine it could well
be useful in future.
